BREAKING CHANGE: Refactors reqstool decorators build hook
