+++ .github/workflows/build.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Install dependencies
         run: pip install hatch
       - name: Run unit and integrations tests
-        run: hatch run test:pytest --cov=reqstool-python-decorators --cov-report=xml --cov-report=html 
+        run: hatch run dev:pytest --cov=reqstool-python-decorators --cov-report=xml --cov-report=html 
       - name: Build project
         run: hatch build
       # Upload artifacts for later use
