+++ .github/workflows/lint.yml
@@ -16,6 +16,6 @@ jobs:
       - name: Install pip package(s)
         run: pip install hatch
       - name: Run black formatter check
-        run: hatch run lint:black --check --verbose src tests
+        run: hatch run dev:black --check --verbose src tests
       - name: Run flake8 linter
-        run: hatch run lint:flake8
+        run: hatch run dev:flake8
