+++ src/reqstool_python_hatch_plugin/hooks.py
@@ -1,9 +1,10 @@
 # Copyright © LFV
 
 from hatchling.plugin import hookimpl
-from reqstool_python_hatch_plugin.build_hook.hook import Decorator
+
+from reqstool_python_hatch_plugin.build_hooks.reqstool_decorators import ReqstoolDecorators
 
 
 @hookimpl
 def hatch_register_build_hook():
-    return Decorator
+    return ReqstoolDecorators
